feat#5: 작성 후 article로  이동 구현
