+++ frontend/toy/src/Home.jsx
@@ -1,12 +1,10 @@
 import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom'
-import axios from 'axios';
 import './Home.css';
 import Header from './layout/Header';
 import Footer from './layout/Footer';
 import BoardList from './board/BoardList';
+import WriteArticle from './board/WriteArticle';
 
-const serverUrl = "http://localhost:8080/article"
 
 /* Home.jsx
     헤더, 게시판 이동 버튼, 푸터로 구성.
@@ -16,49 +14,6 @@ const serverUrl = "http://localhost:8080/article"
 function Home() {
   const [mode, setMode] = useState('Board');
   let mainContent = null;
-  const navigate = useNavigate()
-
-  const [article, setArticle] = useState({
-      title: '',
-      content: ''
-    })
-  
-  const {title, content} = article;
-
-  const config =  {
-    headers: {
-      "Content-Type": "application/json"
-    } 
-  }
-
-  const onChange = (event) => {
-    setArticle({
-        ...article,
-        [event.target.name]: event.target.value,
-    })
-  }
-  
-  const saveArticle = ()  => {
-    axios.post(serverUrl, JSON.stringify(article), config
-    ).then(res =>{
-        if(res.status === 201){
-          alert("작성 완료")
-          setMode('Board')
-          const articleUrl = res.headers.location
-          navigate(articleUrl)
-        }
-        else{
-            alert(res.message)
-        }
-    }).catch((err) => {
-        alert(err.message)
-    })
-  }
-  
-  const handleSubmit = (event) => {
-    event.preventDefault()
-    saveArticle()
-  }
 
   if(mode === 'Board'){
     mainContent = 
@@ -69,11 +24,9 @@ function Home() {
   }
   else if(mode === 'Write'){
     mainContent = 
-      <form onSubmit={handleSubmit}>
-        <p><input type="text" name="title" value={title} onChange={onChange}/></p>
-        <p><textarea name="content" value={content} onChange={onChange}/></p>
-        <input type="submit" value="게시"/>
-      </form>
+    <>
+      <WriteArticle/>
+    </>
   }
 
   return (
