+++ frontend/toy/src/board/Board.jsx
@@ -10,11 +10,9 @@ const Board = ({articleId, title, content, lastDate}) => {
                 <h4>{content}</h4>
             </div>
             <div>
-                <h6>{articleId}</h6>
+                <h6>id : {articleId}</h6>
                 <h6>{lastDate}</h6>
             </div>
-            <hr/>
-            <h2>대충 댓글창</h2>
         </>
     )
 }
